Optimize typed multi-key allocs (StringKeyCodec fast path) and split typed/primeable_typed files to mirror untyped layout
